refactor command folders 📂
